requirements.txt
Django==1.9.2
django-crispy-forms==1.6.0
django-filter==0.12.0
django-filters==0.2.1
djangorestframework==3.3.2
drfdocs==0.0.7
wheel==0.24.0
